requirements.txt
streamlit
msal
msal_streamlit_authentication
load_dotenv